refactor(groups): split group session durable object
- Keep GroupSessionDO focused on RPC, WebSocket, hydration, and broadcast orchestration.
- Move task, rating, pairing, and reset-pool flows into action modules so each mutation can be reviewed independently.
- Move SQLite snapshot helpers, Prisma membership lookup, pairing history, and shared request types into feature-local Durable Object modules.
