+++ src/durable-objects/group-session/runtime.ts
@@ -1,0 +1,8 @@
+import type { PrismaClient } from './types'
+import type { GroupSessionEvent } from '@/features/groups/lib/groupSessionEvents'
+
+export interface GroupSessionRuntime {
+  ctx: DurableObjectState
+  ensureHydrated: (groupId: string, prisma?: PrismaClient) => Promise<void>
+  broadcast: (event: GroupSessionEvent) => void
+}
